Fix circular importer dependency
